Updates to Selenium framework regarding pageLoad and element timeout exceptions. Also quickly added some performance test tools

+++ gurukula-test/src/main/java/com/gurukula/ui/pagemodel/webelements/Element.java
@@ -89,7 +89,7 @@ public class Element {
         return waitForVisible(waitTime);
     }
 
-    public Element waitForPresentAndVisible(int waitTime) {
+    public Element waitForPresentAndVisible(long waitTime) {
         sel.waitForPresentAndVisible(locator, waitTime);
         return this;
     }
